Backend/requirements.txt
# Core Dependencies
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning & Deep Learning
xgboost>=1.7.0
lightgbm>=4.0.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# Data Processing
scipy>=1.10.0
joblib>=1.3.0

# API & Web
requests>=2.31.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
